pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["honjen <144600573+honjen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.6.1"
invoke = "^2.2.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.8.0"
pylint = "^3.3.6"
autopep8 = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
